fix: split legalBasis table in dataset form to 3 seperate tables
