+++ apps/dataset-catalog/components/dataset-form/components/about-section.tsx
@@ -3,7 +3,7 @@ import { localization } from '@catalog-frontend/utils';
 import { Box, Textfield } from '@digdir/designsystemet-react';
 import { FastField } from 'formik';
 import { FieldsetDivider } from '@catalog-frontend/ui';
-import { AccessRightFields } from './access-rights.tsx/access-rights-section';
+import { AccessRightFields } from './access-rights-fields';
 
 export const AboutSection = () => {
   return (
