+++ apps/dataset-catalog/components/dataset-form/components/access-rights-fields.tsx
@@ -3,7 +3,7 @@ import { FieldsetDivider, TitleWithHelpTextAndTag } from '@catalog-frontend/ui';
 import { accessRights, getTranslateText, localization } from '@catalog-frontend/utils';
 import { Box, Radio } from '@digdir/designsystemet-react';
 import { useFormikContext } from 'formik';
-import { AccessRightsUriTable } from './access-rights-uri-table';
+import { UriWithLabelFieldsetTable } from './uri-with-label-field-set-table';
 
 export const AccessRightFields = () => {
   const { values, setFieldValue } = useFormikContext<Dataset>();
@@ -36,8 +36,45 @@ export const AccessRightFields = () => {
           ))}
         </Radio.Group>
       </Box>
+
       <FieldsetDivider />
-      <AccessRightsUriTable />
+
+      <UriWithLabelFieldsetTable
+        values={values.legalBasisForRestriction}
+        fieldName={'legalBasisForRestriction'}
+        hideHeadWhenEmpty={true}
+        label={
+          <TitleWithHelpTextAndTag helpText={localization.datasetForm.helptext.legalBasisForRestriction}>
+            {localization.datasetForm.fieldLabel.legalBasisForRestriction}
+          </TitleWithHelpTextAndTag>
+        }
+      />
+
+      <FieldsetDivider />
+
+      <UriWithLabelFieldsetTable
+        values={values.legalBasisForProcessing}
+        fieldName={'legalBasisForProcessing'}
+        hideHeadWhenEmpty={true}
+        label={
+          <TitleWithHelpTextAndTag helpText={localization.datasetForm.helptext.legalBasisForProcessing}>
+            {localization.datasetForm.fieldLabel.legalBasisForProcessing}
+          </TitleWithHelpTextAndTag>
+        }
+      />
+
+      <FieldsetDivider />
+
+      <UriWithLabelFieldsetTable
+        values={values.legalBasisForAccess}
+        fieldName={'legalBasisForAccess'}
+        hideHeadWhenEmpty={true}
+        label={
+          <TitleWithHelpTextAndTag helpText={localization.datasetForm.helptext.legalBasisForAccess}>
+            {localization.datasetForm.fieldLabel.legalBasisForAccess}
+          </TitleWithHelpTextAndTag>
+        }
+      />
     </>
   );
 };
